[#7] Modify the web server so that it is able to serve static files
The web server needs to be able to server static files requested from
it under the YottaDB path, so that the app can server its static files.
This is needed for production. During development, the static files are
served by nodejs.

this change is so the web server can serve the files without the need
to run npm or node.
in the working directory run quasar build, which generates the static
files in a folder called dist, which then gets served by the web server
directly, paving the road for running the module independently from npm
and node.

to test this specific commit, please refer to "Running the app in producation mode."
section in the README.