[#31] The routines module need to exclude Octo generated routines
The routines module need to exclude Octo generated routines
from showing in the routine list.

This commit adds a button to allow the end user to select whether to include
systems generated routines or not.

The commit also includes a new test in ```tests/routines.spec.js``` to test the feature.